[102] fix visual view
