+++ src/index.html
@@ -49,7 +49,6 @@
       body,
       main {
         width: 100%;
-        height: auto;
       }
 
       @font-face {
@@ -138,6 +137,11 @@
       .logo-position-start img {
         position: relative;
         top: 10px;
+      }
+
+      .footer-logo {
+        width: 69px;
+        margin: 10px 20px 5px 20px;
       }
     </style>
