+++ src/views/visual/visual-view.js
@@ -1,4 +1,4 @@
-import { customElement, html, LitElement, property } from 'lit-element';
+import { customElement, html, css, LitElement, property } from 'lit-element';
 
 import '@google-web-components/google-chart/google-chart.js';
 
@@ -11,6 +11,31 @@ export class VisualView extends LitElement {
   @property({ type: String }) colorScheme;
   @property({ type: Array }) selectedCollections;
 
+  static get styles() {
+    return [
+      css`
+        .visual-view-container {
+          display: flex;
+          flex-direction: column;
+          min-height: 200px;
+          align-items: end;
+        }
+
+        visual-view-header {
+          margin-top: 100px;
+          margin-bottom: 32px;
+        }
+
+        visual-view-graph {
+          margin-bottom: 12px;
+          flex: 1;
+          display: flex;
+          width: 100%;
+        }
+      `,
+    ];
+  }
+
   setSelection = (searchItem, selectedCollections) => {
     this.searchItem = searchItem;
     this.selectedCollections = selectedCollections;
@@ -22,11 +47,12 @@ export class VisualView extends LitElement {
 
   render() {
     return html`
-      <visual-view-header
-        .setSelection="${this.setSelection}"
-        .setColorScheme="${this.setColorScheme}"
-      ></visual-view-header>
-      <div class="graph-wrapper" style="margin-bottom: 12px">
+      <div class="visual-view-container">
+        <visual-view-header
+          .setSelection="${this.setSelection}"
+          .setColorScheme="${this.setColorScheme}"
+        ></visual-view-header>
+
         <visual-view-graph
           .searchItem="${this.searchItem}"
           .colorScheme="${this.colorScheme}"
